update pizza model by sepperating dough and size from the pizza class; moved dough and size to an Order model since these are something you choose on order
