+++ backend/src/main/kotlin/models/order/Dough.kt
@@ -1,4 +1,4 @@
-package models.pizza
+package models.order
 
 /**
  * Lets assume Italian style dough is worth as much as American :)
