+++ backend/src/main/kotlin/models/order/Order.kt
@@ -1,0 +1,5 @@
+package models.order
+
+import models.pizza.Pizza
+
+class Order(val size: Size, val dough: Dough, pizza: Pizza)
