+++ backend/src/main/kotlin/models/pizza/Cheese.kt
@@ -1,0 +1,7 @@
+package models.pizza
+
+enum class Cheese(val value: Int) {
+    MOZERELLA(10),
+    CHEDDAR(10),
+    PARMESAN(10)
+}
