--- backend/src/main/kotlin/models/pizza/PizzaType.kt
@@ -1,5 +1,0 @@
-package models.pizza
-
-enum class PizzaType {
-    HAWAIANO, PEPPERONI, CAPRICIOSA
-}
