+++ backend/src/main/kotlin/models/pizza/Sauce.kt
@@ -1,0 +1,9 @@
+package models.pizza
+
+enum class Sauce {
+    TOMATO, BBQ, TACO;
+
+    override fun toString(): String {
+        return super.toString() + "-SAUCE"
+    }
+}
