+++ backend/src/main/kotlin/models/pizza/Topping.kt
@@ -1,14 +1,16 @@
 package models.pizza
 
 enum class Topping(val value: Int) {
-    BEEF(20),
+    MARINATED_BEEF(30),
+    MARINATED_CHICKEN(30),
+    KEBAB_MEAT(30),
     HAM(10),
-    CHICKEN(20),
     BACON(10),
     PAPRIKA(5),
     CORN(5),
     MUSHROOM(5),
     TORTILLA_CHIPS(5),
-    CHEDDAR(10),
-    MOZERELLA(10),
+    JALEPENOS(5),
+    PEPPERONI(5),
+    ONION(5),
 }
